[symphony+] MIK-6561: agent output
Symphony-Stage: implement
Symphony-Role: implementer/reasonix-deepseek
Symphony-Cli: reasonix
Symphony-Model: deepseek-v4-pro
Symphony-Family: reasonix
Symphony-TickUuid: 3c286d96-2e34-409b-b7e7-c4d57020776e
Symphony-Worker: reasonix-deepseek
Symphony-SpecQuality: high_quality_spec

+++ src/capability/import/generator.rs
@@ -426,7 +426,7 @@ impl ImportGenerator {
         );
         report.insert(
             "total_drafts".to_string(),
-            serde_json::Value::Number(serde_json::Number::from(total)),
+            serde_json::Value::Number(serde_json::Number::from(total as u64)),
         );
         report.insert(
             "pending_review".to_string(),
